+++ src/marker.hpp
@@ -25,20 +25,20 @@
 
 struct Marker {
     Q_GADGET
-    Q_PROPERTY(double position MEMBER position)
+    Q_PROPERTY(int position MEMBER position)
     Q_PROPERTY(int length MEMBER length)
     Q_PROPERTY(QString url MEMBER url)
 public:
     Marker() {
         position = 0;
     };
-    explicit Marker(double p)
+    explicit Marker(int p)
     {
         position = p;
     };
     // Contents
     QString text;
-    double position = 0;
+    int position = 0;
     int length = 1;
     int key = 0;
     QString keyLetter;
